Rename map components with more sensible names
