+++ src/map/Heatmap/animationLoop.js
@@ -6,7 +6,7 @@ import {
   ClipSpaceQuad,
   setParameters
 } from 'luma.gl'
-import ramp from './ramp_plasma.png'
+import ramp from '../colorramp/ramp_plasma.png'
 import fragmentShader from './fragmentShader.glsl'
 import blurV from './blurV.glsl'
 import blurH from './blurH.glsl'
