+++ src/map/RawDataOnRamp/MapRawDataOnRamp.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import animationLoop from './animationLoop'
 
-class ObservationStaticHeatmap extends Component {
+class MapRawDataOnRamp extends Component {
   componentDidMount() {
     this.setProps(this.props)
 //    window.setTimeout(()=>this.stopAnimation(), 5000);
@@ -23,7 +23,6 @@ class ObservationStaticHeatmap extends Component {
     animationLoop.filterMin = nextProps.filterMin
     animationLoop.filterMax = nextProps.filterMax
     animationLoop.alpha = nextProps.alpha
-    animationLoop.amplifyFactor = nextProps.amplifyFactor || 1.
   }
 
   componentWillUnmount() {
@@ -38,9 +37,11 @@ class ObservationStaticHeatmap extends Component {
   }
 }
 
-ObservationStaticHeatmap.propTypes = {
+MapRawDataOnRamp.propTypes = {
+  width: PropTypes.number.isRequired,
+  height: PropTypes.number.isRequired,
   alpha: PropTypes.number,
-  taxonId: PropTypes.number
+  taxonId: PropTypes.string.isRequired
 }
 
-export default ObservationStaticHeatmap
+export default MapRawDataOnRamp
