+++ src/map/colorramp/ramp_plasma.png
